// src/frontend/src/backend.d.ts
export interface Project {
    id: string;
    title: string;
    dateCompleted: bigint;
    description: string;
    imageUrl: string;
    category: ServiceType;
}
export interface UserProfile {
    name: string;
    mobileNumber: string;
    email: string;
}
export enum QuotationStatus {
    rejected = "rejected",
    accepted = "accepted",
    negotiating = "negotiating",
    pendingCustomerResponse = "pendingCustomerResponse"
}
export enum ServiceType {
    banner = "banner",
    offset = "offset",
    design = "design",
    digital = "digital"
}